Moved DefaultDockerGateway to the library
