+++ Mittons.Fixtures.Tests.Integration/Docker/Gateways/DefaultDockerGatewayTests.cs
@@ -1,6 +1,5 @@
 using Xunit;
 using Mittons.Fixtures.Docker.Gateways;
-using Mittons.Fixtures.Docker.Containers;
 using System.Diagnostics;
 using System.Text;
 using System;
@@ -8,34 +7,6 @@ using System.Collections.Generic;
 
 namespace Mittons.Fixtures.Tests.Integration.Docker.Gateways
 {
-    public class DefaultDockerGateway : IDockerGateway
-    {
-        public void Remove(string containerName)
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public string Run(string imageName, string command)
-        {
-            var proc = new Process();
-            proc.StartInfo.FileName = "docker";
-            proc.StartInfo.Arguments = $"run -d {imageName} {command}";
-            proc.StartInfo.UseShellExecute = false;
-            proc.StartInfo.RedirectStandardOutput = true;
-
-            proc.Start();
-            proc.WaitForExit();
-
-            var containerId = default(string);
-
-            while (!proc.StandardOutput.EndOfStream) {
-                containerId = proc.StandardOutput.ReadLine();
-            }
-
-            return containerId ?? string.Empty;
-        }
-    }
-
     public class DefaultDockerGatewayTests : IDisposable
     {
         private List<string> containerIds = new List<string>();
